Backend/requirements.txt
fastapi
uvicorn
pymongo
python-multipart
pydantic
passlib
python-jose[cryptography]